refactor: extract useBlackboxViewer composable; start Pinia bridge collapse (#924 Phase 1)
Establish the Phase 1 architecture (composable owns renderer instances +
imperative ops; Pinia stores hold pure state):

- git mv main.js -> composables/use_blackbox_viewer.js (body verbatim),
  wrapped as a singleton `useBlackboxViewer()` that initializes once and
  returns its operations. Relative imports fixed for the new location.
- main.js is now a thin bootstrap that calls useBlackboxViewer().
- First bridge group collapsed end-to-end as the pattern: spectrum
  export/import/clear are exposed by the composable (not graphStore
  shallowRef callbacks); removed the 3 refs from graph store;
  SpectrumAnalyser.vue now calls the composable directly.

Build, type-check, lint, and the parser golden test pass. NOTE: this
rewires app bootstrap — a green build does not prove runtime; behavior
should be identical (relocated code) but needs an app run to confirm.
Remaining ~42 store callback refs collapse in follow-up increments.

Refs #924
