+++ src/components/SpectrumAnalyser.vue
@@ -93,8 +93,10 @@ import { ref, computed, watch } from "vue";
 import { useGraphStore } from "../stores/graph.js";
 import { useSettingsStore } from "../stores/settings.js";
 import { SPECTRUM_TYPE } from "../graph_spectrum_plot";
+import { useBlackboxViewer } from "../composables/use_blackbox_viewer.js";
 
 const graphStore = useGraphStore();
+const viewer = useBlackboxViewer();
 const { userSettings } = useSettingsStore();
 
 const importInput = ref(null);
@@ -225,14 +227,14 @@ function resetSegmentLength() { segmentLength.value = 9; getAnalyser()?.resetSeg
 // --- Spectrum actions ---
 const spectrumMenuItems = [
   [
-    { label: "Export CSV", icon: "i-lucide-download", onSelect: () => graphStore.spectrumExport?.() },
+    { label: "Export CSV", icon: "i-lucide-download", onSelect: () => viewer.spectrumExport() },
     { label: "Import CSV", icon: "i-lucide-upload", onSelect: () => importInput.value?.click() },
-    { label: "Clear imported", icon: "i-lucide-trash-2", onSelect: () => graphStore.spectrumClear?.() },
+    { label: "Clear imported", icon: "i-lucide-trash-2", onSelect: () => viewer.spectrumClear() },
   ],
 ];
 
 function onImportChange(e) {
-  graphStore.spectrumImport?.(e.target.files);
+  viewer.spectrumImport(e.target.files);
   e.target.value = "";
 }
