+++ src/stores/graph.js
@@ -64,9 +64,6 @@ export const useGraphStore = defineStore("graph", () => {
   const reorderGraphs = shallowRef(null);
   const resetPen = shallowRef(null);
   const fieldWheel = shallowRef(null);
-  const spectrumExport = shallowRef(null);
-  const spectrumImport = shallowRef(null);
-  const spectrumClear = shallowRef(null);
   const applyGraphZoom = shallowRef(null);
   const selectLogIndex = shallowRef(null);
   const setSeekBarMode = shallowRef(null);
@@ -218,9 +215,6 @@ export const useGraphStore = defineStore("graph", () => {
     reorderGraphs,
     resetPen,
     fieldWheel,
-    spectrumExport,
-    spectrumImport,
-    spectrumClear,
     applyGraphZoom,
     selectLogIndex,
     setSeekBarMode,
